Bot can check price
